# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Project Timesheet API Documentation",
        "description": "API Authentication for User Project Timesheet by Delfrinando",
        "contact": {
            "email": "delfrinando@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/timesheets": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get list of timesheets",
                "description": "Get a list of timesheets for the authenticated user",
                "operationId": "getTimesheetsList",
                "responses": {
                    "200": {
                        "description": "Timesheets fetched successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Timesheets not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Create a new timesheet",
                "description": "Create a new timesheet and assign it to a project",
                "operationId": "createTimesheet",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task_name": {
                                        "type": "string",
                                        "example": "Task Alpha"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.5
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Timesheet created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/timesheets/{id}": {
            "get": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Get details of a specific timesheet",
                "description": "Fetches the details of a specific timesheet by its ID",
                "operationId": "getTimesheetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet details fetched successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Update an existing timesheet",
                "description": "Update the details of an existing timesheet by its ID",
                "operationId": "updateTimesheetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "task_name": {
                                        "type": "string",
                                        "example": "Updated Task"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-02-01"
                                    },
                                    "hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 7.5
                                    },
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Timesheet updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Timesheets"
                ],
                "summary": "Delete a timesheet",
                "description": "Delete a timesheet by its ID",
                "operationId": "deleteTimesheetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timesheet ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Timesheet deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Timesheet not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}